requirements.txt
# Core FastAPI & Web
fastapi==0.111.*
uvicorn[standard]==0.29.*
pydantic==2.7.*
python-dotenv==1.0.*
PyYAML==6.0.*
httpx==0.27.*
requests==2.31.* # requests might still be used by libraries

# LangChain & RAG
langchain==0.1.*
langchain-community==0.0.* # Check latest compatible versions
faiss-cpu==1.8.* # Or faiss-gpu if you have CUDA setup
sentence-transformers==2.7.*
# If using LangChain's Cohere integration:
langchain-cohere==0.1.*

# NLP & Utilities
nltk==3.8.*
spacy==3.7.*
# Download spacy model separately: python -m spacy download en_core_web_sm
transformers==4.41.*
torch==2.2.* # Or torch appropriate for your system (CPU/CUDA/MPS)
networkx==3.3.*
pandas==2.2.* # For CSV ingestion

# Caching
fastapi-cache2[redis]==0.2.*
redis

# Search API SDK (Example - install the one you chose)
# google-api-python-client # If using Google Custom Search JSON API
# searchapi-python # If using SearchApi.io Python SDK

# Cohere SDK (if calling directly, optional if langchain-cohere works)
cohere==5.3.*
